chore: refactored into better folder structure
